Use ZFS build-in fio-config
This Commit makes sure we use the fio config that is already build into zfs.
We still use (most of) our own fixes when it comes to compression (repeatability),
but those are added to the fio config thats already in ZFS.

However sequential read-write is not included in zfs and is supplied seperately.

Both the sequential read-write config and the config tweaks are also send to ZFSonLinux in a PR:
https://github.com/zfsonlinux/zfs/pull/9793